+++ tests/Predis/Connection/PhpiredisStreamConnectionTest.php
@@ -73,6 +73,24 @@ class PhpiredisStreamConnectionTest extends ConnectionTestCase
     /**
      * @group connected
      */
+    public function testAcceptsTcpNodelayParameter()
+    {
+        if (!version_compare(PHP_VERSION, '5.4.0', '>=')) {
+            $this->markTestSkipped('Setting TCP_NODELAY on PHP socket streams works on PHP >= 5.4.0');
+        }
+
+        $connection = new PhpiredisStreamConnection($this->getParameters(array('tcp_nodelay' => false)));
+        $connection->connect();
+        $this->assertTrue($connection->isConnected());
+
+        $connection = new PhpiredisStreamConnection($this->getParameters(array('tcp_nodelay' => true)));
+        $connection->connect();
+        $this->assertTrue($connection->isConnected());
+    }
+
+    /**
+     * @group connected
+     */
     public function testExecutesCommandsOnServer()
     {
         $connection = $this->getConnection($profile, true);