+++ tests/Predis/Connection/StreamConnectionTest.php
@@ -72,6 +72,24 @@ class StreamConnectionTest extends ConnectionTestCase
     /**
      * @group connected
      */
+    public function testAcceptsTcpNodelayParameter()
+    {
+        if (!version_compare(PHP_VERSION, '5.4.0', '>=')) {
+            $this->markTestSkipped('Setting TCP_NODELAY on PHP socket streams works on PHP >= 5.4.0');
+        }
+
+        $connection = new StreamConnection($this->getParameters(array('tcp_nodelay' => false)));
+        $connection->connect();
+        $this->assertTrue($connection->isConnected());
+
+        $connection = new StreamConnection($this->getParameters(array('tcp_nodelay' => true)));
+        $connection->connect();
+        $this->assertTrue($connection->isConnected());
+    }
+
+    /**
+     * @group connected
+     */
     public function testReadsMultibulkRepliesAsIterators()
     {
         $connection = $this->getConnection($profile, true, array('iterable_multibulk' => true));